+++ OpenGLProject/OpenGLProject/Camera.cpp
@@ -74,7 +74,7 @@ void Camera::UpdateCameraVectors()
 
 void Camera::AllocateProjectionView(const std::vector<uint32_t>& shaderIDs)
 {
-	ubo = new UniformBuffer(shaderIDs, "ubo_ProjectionView", static_cast<size_t>(Utils::mat4v4Size));
+	ubo = new UniformBuffer(shaderIDs, "ubo_ProjectionView", Utils::mat4v4Size);
 }
 
 void Camera::SetProjectionViewUniform(const float windowAspectRatio)
