+++ OpenGLProject/OpenGLProject/Material.cpp
@@ -15,7 +15,7 @@ Material::Material(const glm::vec3& inSpecular, const float inShininess) :
 
 void Material::Store(const std::vector<uint32_t>& entitiesShadersIDs)
 {
-	ubo = new UniformBuffer(entitiesShadersIDs, "specularMatParams", static_cast<size_t>(Utils::vec4Size + Utils::scalarSize));
+	ubo = new UniformBuffer(entitiesShadersIDs, "specularMatParams", Utils::vec4Size + Utils::scalarSize);
 	ubo->InitSubData({
 		{ static_cast<const void*>(glm::value_ptr(specularColour)), Utils::vec4Size },
 		{ static_cast<const void*>(&shininess), Utils::scalarSize }
